# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_listAuditLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回审计日志列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "操作审计日志",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "operationId": "AdminController_getDashboardStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回统计数据"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "管理后台数据概览",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/admin/identity/pending": {
      "get": {
        "operationId": "AdminController_listPendingIdentities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回待审核实名列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "待审核实名列表",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/identity/{id}/approve": {
      "post": {
        "operationId": "AdminController_approveIdentity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "认证通过"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "通过实名认证",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/identity/{id}/reject": {
      "post": {
        "operationId": "AdminController_rejectIdentity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectIdentityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已拒绝"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "拒绝实名认证",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/admin/login-logs": {
      "get": {
        "operationId": "AdminController_listLoginLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回登录日志列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "登录日志",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/merchants": {
      "get": {
        "description": "分页查询所有商户",
        "operationId": "AdminController_listMerchants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回商户列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "商户列表",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/merchants/{id}/audit": {
      "post": {
        "description": "通过或拒绝商户入驻申请",
        "operationId": "AdminController_auditMerchant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditMerchantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "审核完成"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "审核商户",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/merchants/{id}/config": {
      "post": {
        "description": "修改费率、日限额等商户配置",
        "operationId": "AdminController_updateMerchantConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMerchantConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "配置更新成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "修改商户配置",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/payment-orders": {
      "get": {
        "operationId": "AdminController_listPaymentOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回支付订单列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "支付订单列表",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/admin/risk-events": {
      "get": {
        "operationId": "AdminController_listRiskEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回风控事件列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "风控事件列表",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/risk-events/{id}/handle": {
      "post": {
        "operationId": "AdminController_handleRiskEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleRiskEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "事件已处理"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "处理风控事件",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/risk-rules": {
      "get": {
        "operationId": "AdminController_getRiskRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回风控规则列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "获取风控规则",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/risk-rules/{code}": {
      "put": {
        "operationId": "AdminController_updateRiskRule",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRiskRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "规则更新成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "更新风控规则",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "description": "分页查询所有用户",
        "operationId": "AdminController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回用户列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "用户列表",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回用户详情"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "用户详情",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/users/{id}/risk-level": {
      "post": {
        "operationId": "AdminController_updateUserRiskLevel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRiskLevelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "风控等级更新成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "修改用户风控等级",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/users/{id}/status": {
      "post": {
        "description": "冻结/解冻用户账号",
        "operationId": "AdminController_updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "状态更新成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "修改用户状态",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/merchants/dashboard": {
      "get": {
        "description": "返回今日交易额、订单数等统计",
        "operationId": "MerchantsController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回统计数据"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "商户数据看板",
        "tags": [
          "商户"
        ]
      }
    },
...
    "/users/login-logs": {
      "get": {
        "description": "获取当前用户最近 30 天登录记录",
        "operationId": "UsersController_loginLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回登录日志列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "查询登录日志",
        "tags": [
          "用户"
        ]
      }
    },
...
  },
  "info": {
    "title": "科佰支付 KeBaiPay API",
    "description": "合规聚合支付技术服务商接口文档：资金绝不过平台，仅做收单、退款与对账聚合。\n\n## 认证方式\n- **用户接口**：使用 JWT Bearer Token 认证\n- **商户开放 API**：使用 HMAC-SHA256 签名认证\n- **管理后台**：使用管理员 JWT Token 认证\n\n## 错误码规范\n所有错误返回统一格式：`{ \"statusCode\": number, \"message\": \"KBxxx 错误描述\" }`\n\n## 金额说明\n数据库存储单位为「分」，接口传输单位为「元」\n\n> 已下线接口（v0.3.x 精简）：多平台对账聚合（/admin/channel-reconciliation/*）、AI 风控审计（/risk-audit/*、/admin/risk-audit/*）、自定义风控规则（/admin/risk-rules/custom、/risk-rules/custom）。对账以 /admin/reconciliation 为准。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "用户/管理员 JWT Token"
      },
      "app-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-App-Id",
        "description": "商户应用 App ID"
      }
    },
    "schemas": {
...
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserRiskLevelDto": {
        "type": "object",
        "properties": {}
      },
      "AuditMerchantDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMerchantConfigDto": {
        "type": "object",
        "properties": {}
      },
...
      "HandleRiskEventDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateRiskRuleDto": {
        "type": "object",
        "properties": {}
      },
      "RejectIdentityDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}